Final da refatoração
