+++ semana20/Labook/src/index.ts
@@ -1,93 +1,6 @@
-import { Request, Response } from "express";
-import { authenticationData } from "./types/authenticationData";
-import { user } from "./types/user";
-import { post } from "./types/post";
 import app from "./app";
-import { connection } from "./connection";
-import { IdGenerator } from "./services/IdGenerator";
-import { HashManager } from "./services/HashManager";
-import { Authenticator } from "./services/Authenticator";
 import { postsRouter } from "./routes/postRouter";
-import { createPostController } from "./controller/post/createPostController";
-import { selectUserByEmail } from "./data/user/selectUserByEmail";
-import { insertUser } from "./data/user/insertUser";
+import { usersRouter } from "./routes/userRoutes";
 
-const authenticator = new Authenticator();
-const hashManager = new HashManager();
-const idGenerator = new IdGenerator();
-
-/**************************** ENDPOINTS ******************************/
-
-
-
-app.post("/users/login", async (req: Request, res: Response) => {
-  try {
-    let message = "Success!";
-
-    const { email, password } = req.body;
-
-    if (!email || !password) {
-      res.statusCode = 406;
-      message = '"email" and "password" must be provided';
-      throw new Error(message);
-    }
-
-    const user = await selectUserByEmail(email);
-
-    const passwordIsCorrect: boolean = await hashManager.compare(
-      password,
-      user.password
-    );
-
-    if (!passwordIsCorrect) {
-      res.statusCode = 401;
-      message = "Invalid credentials";
-      throw new Error(message);
-    }
-
-    const token: string = authenticator.generateToken({
-      id: user.id,
-    });
-
-    res.status(200).send({ message, token });
-  } catch (error) {
-    let message = error.sqlMessage || error.message;
-    res.statusCode = 400;
-
-    res.send({ message });
-  }
-});
-
-app.get("/posts/:id", async (req: Request, res: Response) => {
-  try {
-    let message = "Success!";
-
-    const { id } = req.params;
-
-    const queryResult: any = await connection("labook_posts")
-      .select("*")
-      .where({ id });
-
-    if (!queryResult[0]) {
-      res.statusCode = 404;
-      message = "Post not found";
-      throw new Error(message);
-    }
-
-    const post: post = {
-      id: queryResult[0].id,
-      photo: queryResult[0].photo,
-      description: queryResult[0].description,
-      type: queryResult[0].type,
-      createdAt: queryResult[0].created_at,
-      authorId: queryResult[0].author_id,
-    };
-
-    res.status(200).send({ message, post });
-  } catch (error) {
-    let message = error.sqlMessage || error.message;
-    res.statusCode = 400;
-
-    res.send({ message });
-  }
-});
+app.use("/users", usersRouter);
+app.use("/posts", postsRouter);
